Ensure PCA doesn't replace input data (#9936)
* Ensure PCA doesn't replace input data

* Duplicate A within pca_factor instead.

+++ nd4j/nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/blas/impl/BaseLapack.java
@@ -218,8 +218,8 @@ public abstract class BaseLapack implements Lapack {
         if (A.rows() > Integer.MAX_VALUE || A.columns() > Integer.MAX_VALUE)
             throw new ND4JArraySizeException();
 
-        int m = (int) A.rows();
-        int n = (int) A.columns();
+        int m = A.rows();
+        int n = A.columns();
 
         byte jobu = (byte) (U == null ? 'N' : 'A');
         byte jobvt = (byte) (VT == null ? 'N' : 'A');
